src/services/ctf/requirements.txt
APScheduler==3.10.4
black==24.10.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docker==7.1.0
durationpy==0.9
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-APScheduler==1.13.1
google-auth==2.35.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==31.0.0
MarkupSafe==3.0.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
#pywin32==308
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
testcontainers==4.8.2
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
waitress==3.0.0
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
